+++ pkg/bpmn_engine/conditions_test.go
@@ -13,8 +13,8 @@ func Test_exclusive_gateway_with_expressions_selects_one_and_not_the_other(t *te
 
 	// given
 	process, _ := bpmnEngine.LoadFromFile("../../test-cases/exclusive-gateway-with-condition.bpmn")
-	bpmnEngine.NewTaskHandler().Id("task-a").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().Id("task-b").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Id("task-a").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Id("task-b").Handler(cp.TaskHandler)
 	variables := map[string]interface{}{
 		"price": -50,
 	}
@@ -33,8 +33,8 @@ func Test_exclusive_gateway_with_expressions_selects_default(t *testing.T) {
 
 	// given
 	process, _ := bpmnEngine.LoadFromFile("../../test-cases/exclusive-gateway-with-condition-and-default.bpmn")
-	bpmnEngine.NewTaskHandler().Id("task-a").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().Id("task-b").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Id("task-a").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Id("task-b").Handler(cp.TaskHandler)
 	variables := map[string]interface{}{
 		"price": -1,
 	}
@@ -90,8 +90,9 @@ func Test_evaluation_error_percolates_up(t *testing.T) {
 
 	// when
 	// don't provide variables, for execution
-	_, err := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
+	instance, err := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
 
 	// then
+	then.AssertThat(t, instance.State, is.EqualTo(Failed))
 	then.AssertThat(t, err, is.Not(is.Nil()))
 }