+++ pkg/bpmn_engine/engine_fork_join_test.go
@@ -13,9 +13,9 @@ func TestForkUncontrolledJoin(t *testing.T) {
 
 	// given
 	process, _ := bpmnEngine.LoadFromFile("../../test-cases/fork-uncontrolled-join.bpmn")
-	bpmnEngine.NewTaskHandler().Id("id-a-1").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().Id("id-a-2").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().Id("id-b-1").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Id("id-a-1").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Id("id-a-2").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Id("id-b-1").Handler(cp.TaskHandler)
 
 	// when
 	bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
@@ -31,12 +31,13 @@ func TestForkControlledParallelJoin(t *testing.T) {
 
 	// given
 	process, _ := bpmnEngine.LoadFromFile("../../test-cases/fork-controlled-parallel-join.bpmn")
-	bpmnEngine.NewTaskHandler().Id("id-a-1").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().Id("id-a-2").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().Id("id-b-1").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Id("id-a-1").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Id("id-a-2").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Id("id-b-1").Handler(cp.TaskHandler)
 
 	// when
-	bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
+	_, err := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
+	then.AssertThat(t, err, is.Nil())
 
 	// then
 	then.AssertThat(t, cp.CallPath, is.EqualTo("id-a-1,id-a-2,id-b-1"))
@@ -49,9 +50,9 @@ func TestForkControlledExclusiveJoin(t *testing.T) {
 
 	// given
 	process, _ := bpmnEngine.LoadFromFile("../../test-cases/fork-controlled-exclusive-join.bpmn")
-	bpmnEngine.NewTaskHandler().Id("id-a-1").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().Id("id-a-2").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().Id("id-b-1").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Id("id-a-1").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Id("id-a-2").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Id("id-b-1").Handler(cp.TaskHandler)
 
 	// when
 	bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)