+++ pkg/bpmn_engine/events_test.go
@@ -4,8 +4,6 @@ import (
 	"github.com/corbym/gocrest/has"
 	"github.com/corbym/gocrest/is"
 	"github.com/corbym/gocrest/then"
-	"github.com/nitram509/lib-bpmn-engine/pkg/spec/BPMN20/activity"
-	"github.com/nitram509/lib-bpmn-engine/pkg/spec/BPMN20/process_instance"
 	"testing"
 )
 
@@ -19,7 +17,7 @@ func Test_creating_a_process_sets_state_to_READY(t *testing.T) {
 	// when
 	pi, _ := bpmnEngine.CreateInstance(process.ProcessKey, nil)
 	// then
-	then.AssertThat(t, pi.GetState(), is.EqualTo(process_instance.READY))
+	then.AssertThat(t, pi.GetState(), is.EqualTo(Ready))
 }
 
 func Test_running_a_process_sets_state_to_ACTIVE(t *testing.T) {
@@ -34,9 +32,9 @@ func Test_running_a_process_sets_state_to_ACTIVE(t *testing.T) {
 	procInst, _ := bpmnEngine.RunOrContinueInstance(pi.GetInstanceKey())
 
 	// then
-	then.AssertThat(t, pi.GetState(), is.EqualTo(process_instance.ACTIVE).
+	then.AssertThat(t, pi.GetState(), is.EqualTo(Active).
 		Reason("Since the BPMN contains an intermediate catch event, the process instance must be active and can't complete."))
-	then.AssertThat(t, procInst.GetState(), is.EqualTo(process_instance.ACTIVE))
+	then.AssertThat(t, procInst.GetState(), is.EqualTo(Active))
 }
 
 func Test_IntermediateCatchEvent_received_message_completes_the_instance(t *testing.T) {
@@ -48,11 +46,13 @@ func Test_IntermediateCatchEvent_received_message_completes_the_instance(t *test
 	pi, _ := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
 
 	// when
-	bpmnEngine.PublishEventForInstance(pi.GetInstanceKey(), "globalMsgRef", nil)
-	bpmnEngine.RunOrContinueInstance(pi.GetInstanceKey())
+	err := bpmnEngine.PublishEventForInstance(pi.GetInstanceKey(), "globalMsgRef", nil)
+	then.AssertThat(t, err, is.Nil())
+	_, err = bpmnEngine.RunOrContinueInstance(pi.GetInstanceKey())
+	then.AssertThat(t, err, is.Nil())
 
 	// then
-	then.AssertThat(t, pi.GetState(), is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, pi.GetState(), is.EqualTo(Completed))
 }
 
 func Test_IntermediateCatchEvent_message_can_be_published_before_running_the_instance(t *testing.T) {
@@ -68,7 +68,7 @@ func Test_IntermediateCatchEvent_message_can_be_published_before_running_the_ins
 	bpmnEngine.RunOrContinueInstance(pi.GetInstanceKey())
 
 	// then
-	then.AssertThat(t, pi.GetState(), is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, pi.GetState(), is.EqualTo(Completed))
 }
 
 func Test_IntermediateCatchEvent_a_catch_event_produces_an_active_subscription(t *testing.T) {
@@ -85,7 +85,7 @@ func Test_IntermediateCatchEvent_a_catch_event_produces_an_active_subscription(t
 	subscription := subscriptions[0]
 	then.AssertThat(t, subscription.Name, is.EqualTo("event-1"))
 	then.AssertThat(t, subscription.ElementId, is.EqualTo("id-1"))
-	then.AssertThat(t, subscription.State, is.EqualTo(activity.Active))
+	then.AssertThat(t, subscription.State, is.EqualTo(Active))
 }
 
 func Test_Having_IntermediateCatchEvent_and_ServiceTask_in_parallel_the_process_state_is_maintained(t *testing.T) {
@@ -96,14 +96,14 @@ func Test_Having_IntermediateCatchEvent_and_ServiceTask_in_parallel_the_process_
 	// given
 	process, _ := bpmnEngine.LoadFromFile("../../test-cases/message-intermediate-catch-event-and-parallel-tasks.bpmn")
 	instance, _ := bpmnEngine.CreateInstance(process.ProcessKey, nil)
-	bpmnEngine.NewTaskHandler().Id("task-1").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().Id("task-2").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Id("task-1").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Id("task-2").Handler(cp.TaskHandler)
 
 	// when
 	bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
 
 	// then
-	then.AssertThat(t, instance.GetState(), is.EqualTo(process_instance.ACTIVE))
+	then.AssertThat(t, instance.GetState(), is.EqualTo(Active))
 
 	// when
 	bpmnEngine.PublishEventForInstance(instance.GetInstanceKey(), "event-1", nil)
@@ -111,7 +111,7 @@ func Test_Having_IntermediateCatchEvent_and_ServiceTask_in_parallel_the_process_
 
 	// then
 	then.AssertThat(t, cp.CallPath, is.EqualTo("task-2,task-1"))
-	then.AssertThat(t, instance.GetState(), is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, instance.GetState(), is.EqualTo(Completed))
 }
 
 func Test_multiple_intermediate_catch_events_possible(t *testing.T) {
@@ -121,21 +121,22 @@ func Test_multiple_intermediate_catch_events_possible(t *testing.T) {
 
 	// given
 	process, _ := bpmnEngine.LoadFromFile("../../test-cases/message-multiple-intermediate-catch-events.bpmn")
-	bpmnEngine.NewTaskHandler().Id("task1").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().Id("task2").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().Id("task3").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Id("task1").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Id("task2").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Id("task3").Handler(cp.TaskHandler)
 	instance, err := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
 	then.AssertThat(t, err, is.Nil())
 
 	// when
 	err = bpmnEngine.PublishEventForInstance(instance.GetInstanceKey(), "msg-event-2", nil)
 	then.AssertThat(t, err, is.Nil())
-	bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
+	_, err = bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
+	then.AssertThat(t, err, is.Nil())
 
 	// then
 	then.AssertThat(t, cp.CallPath, is.EqualTo("task2"))
 	// then still active, since there's an implicit fork
-	then.AssertThat(t, instance.GetState(), is.EqualTo(process_instance.ACTIVE))
+	then.AssertThat(t, instance.GetState(), is.EqualTo(Active))
 }
 
 func Test_multiple_intermediate_catch_events_implicit_fork_and_merged_COMPLETED(t *testing.T) {
@@ -155,7 +156,7 @@ func Test_multiple_intermediate_catch_events_implicit_fork_and_merged_COMPLETED(
 	bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
 
 	// then
-	then.AssertThat(t, instance.GetState(), is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, instance.GetState(), is.EqualTo(Completed))
 }
 
 func Test_multiple_intermediate_catch_events_implicit_fork_and_merged_ACTIVE(t *testing.T) {
@@ -173,7 +174,7 @@ func Test_multiple_intermediate_catch_events_implicit_fork_and_merged_ACTIVE(t *
 	bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
 
 	// then
-	then.AssertThat(t, instance.GetState(), is.EqualTo(process_instance.ACTIVE))
+	then.AssertThat(t, instance.GetState(), is.EqualTo(Active))
 }
 
 func Test_multiple_intermediate_catch_events_implicit_fork_and_parallel_gateway_COMPLETED(t *testing.T) {
@@ -193,7 +194,7 @@ func Test_multiple_intermediate_catch_events_implicit_fork_and_parallel_gateway_
 	bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
 
 	// then
-	then.AssertThat(t, instance.GetState(), is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, instance.GetState(), is.EqualTo(Completed))
 }
 
 func Test_multiple_intermediate_catch_events_implicit_fork_and_parallel_gateway_ACTIVE(t *testing.T) {
@@ -211,7 +212,7 @@ func Test_multiple_intermediate_catch_events_implicit_fork_and_parallel_gateway_
 	bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
 
 	// then
-	then.AssertThat(t, instance.GetState(), is.EqualTo(process_instance.ACTIVE))
+	then.AssertThat(t, instance.GetState(), is.EqualTo(Active))
 }
 func Test_multiple_intermediate_catch_events_implicit_fork_and_exclusive_gateway_COMPLETED(t *testing.T) {
 	// setup
@@ -230,7 +231,7 @@ func Test_multiple_intermediate_catch_events_implicit_fork_and_exclusive_gateway
 	bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
 
 	// then
-	then.AssertThat(t, instance.GetState(), is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, instance.GetState(), is.EqualTo(Completed))
 }
 
 func Test_multiple_intermediate_catch_events_implicit_fork_and_exclusive_gateway_ACTIVE(t *testing.T) {
@@ -248,7 +249,7 @@ func Test_multiple_intermediate_catch_events_implicit_fork_and_exclusive_gateway
 	bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
 
 	// then
-	then.AssertThat(t, instance.GetState(), is.EqualTo(process_instance.ACTIVE))
+	then.AssertThat(t, instance.GetState(), is.EqualTo(Active))
 }
 
 func Test_publishing_a_random_message_does_no_harm(t *testing.T) {
@@ -267,7 +268,7 @@ func Test_publishing_a_random_message_does_no_harm(t *testing.T) {
 
 	// then
 	then.AssertThat(t, err, is.Nil())
-	then.AssertThat(t, instance.GetState(), is.EqualTo(process_instance.ACTIVE))
+	then.AssertThat(t, instance.GetState(), is.EqualTo(Active))
 }
 
 func Test_eventBasedGateway_just_fires_one_event_and_instance_COMPLETED(t *testing.T) {
@@ -278,8 +279,8 @@ func Test_eventBasedGateway_just_fires_one_event_and_instance_COMPLETED(t *testi
 	// given
 	process, _ := bpmnEngine.LoadFromFile("../../test-cases/message-EventBasedGateway.bpmn")
 	instance, _ := bpmnEngine.CreateInstance(process.ProcessKey, nil)
-	bpmnEngine.NewTaskHandler().Id("task-a").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().Id("task-b").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Id("task-a").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Id("task-b").Handler(cp.TaskHandler)
 
 	// when
 	_ = bpmnEngine.PublishEventForInstance(instance.GetInstanceKey(), "msg-b", nil)
@@ -287,7 +288,7 @@ func Test_eventBasedGateway_just_fires_one_event_and_instance_COMPLETED(t *testi
 
 	// then
 	then.AssertThat(t, cp.CallPath, is.EqualTo("task-b"))
-	then.AssertThat(t, instance.GetState(), is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, instance.GetState(), is.EqualTo(Completed))
 }
 
 func Test_intermediate_message_catch_event_publishes_variables_into_instance(t *testing.T) {
@@ -304,7 +305,7 @@ func Test_intermediate_message_catch_event_publishes_variables_into_instance(t *
 	_, _ = bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
 
 	// then
-	then.AssertThat(t, instance.GetState(), is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, instance.GetState(), is.EqualTo(Completed))
 	then.AssertThat(t, instance.GetVariable("foo"), is.EqualTo("bar"))
 	then.AssertThat(t, instance.GetVariable("mappedFoo"), is.EqualTo("bar"))
 }
@@ -322,7 +323,7 @@ func Test_intermediate_message_catch_event_output_mapping_failed(t *testing.T) {
 	_, _ = bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
 
 	// then
-	then.AssertThat(t, instance.GetState(), is.EqualTo(process_instance.FAILED))
+	then.AssertThat(t, instance.GetState(), is.EqualTo(Failed))
 	then.AssertThat(t, instance.GetVariable("mappedFoo"), is.Nil())
-	then.AssertThat(t, bpmnEngine.messageSubscriptions[0].State, is.EqualTo(activity.Failed))
+	then.AssertThat(t, bpmnEngine.messageSubscriptions[0].State, is.EqualTo(Failed))
 }