+++ pkg/bpmn_engine/timer_test.go
@@ -15,8 +15,8 @@ func TestEventBasedGatewaySelectsPathWhereTimerOccurs(t *testing.T) {
 
 	// given
 	process, _ := bpmnEngine.LoadFromFile("../../test-cases/message-intermediate-timer-event.bpmn")
-	bpmnEngine.NewTaskHandler().Id("task-for-message").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().Id("task-for-timer").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Id("task-for-message").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Id("task-for-timer").Handler(cp.TaskHandler)
 	instance, _ := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
 
 	// when
@@ -34,51 +34,56 @@ func TestInvalidTimer_will_stop_continue_execution(t *testing.T) {
 
 	// given
 	process, _ := bpmnEngine.LoadFromFile("../../test-cases/message-intermediate-invalid-timer-event.bpmn")
-	bpmnEngine.NewTaskHandler().Id("task-for-timer").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Id("task-for-timer").Handler(cp.TaskHandler)
 	instance, _ := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
 
 	// when
-	bpmnEngine.PublishEventForInstance(instance.GetInstanceKey(), "message", nil)
-	bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
+	err := bpmnEngine.PublishEventForInstance(instance.GetInstanceKey(), "message", nil)
+	then.AssertThat(t, err, is.Nil())
+	_, err = bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
+	then.AssertThat(t, err, is.Nil())
 
 	// then
 	then.AssertThat(t, cp.CallPath, is.EqualTo(""))
 }
 
-func TestEventBasedGatewaySelectsPathWhereMessageReceived(t *testing.T) {
+func Test_EventBasedGateway_selects_path_where_message_received(t *testing.T) {
 	// setup
 	bpmnEngine := New()
 	cp := CallPath{}
 
 	// given
 	process, _ := bpmnEngine.LoadFromFile("../../test-cases/message-intermediate-timer-event.bpmn")
-	bpmnEngine.NewTaskHandler().Id("task-for-message").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().Id("task-for-timer").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Id("task-for-message").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Id("task-for-timer").Handler(cp.TaskHandler)
 	instance, _ := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
 
 	// when
 	time.Sleep((1 * time.Second) + (1 * time.Millisecond))
-	bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
+	_, err := bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
+	then.AssertThat(t, err, is.Nil())
 
 	// then
 	then.AssertThat(t, cp.CallPath, is.EqualTo("task-for-timer"))
 }
 
-func TestEventBasedGatewaySelectsJustOnePath(t *testing.T) {
+func Test_EventBasedGateway_selects_just_one_path(t *testing.T) {
 	// setup
 	bpmnEngine := New()
 	cp := CallPath{}
 
 	// given
 	process, _ := bpmnEngine.LoadFromFile("../../test-cases/message-intermediate-timer-event.bpmn")
-	bpmnEngine.NewTaskHandler().Id("task-for-message").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().Id("task-for-timer").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Id("task-for-message").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Id("task-for-timer").Handler(cp.TaskHandler)
 	instance, _ := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
 
 	// when
 	time.Sleep((1 * time.Second) + (1 * time.Millisecond))
-	bpmnEngine.PublishEventForInstance(instance.GetInstanceKey(), "message", nil)
-	bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
+	err := bpmnEngine.PublishEventForInstance(instance.GetInstanceKey(), "message", nil)
+	then.AssertThat(t, err, is.Nil())
+	_, err = bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
+	then.AssertThat(t, err, is.Nil())
 
 	// then
 	then.AssertThat(t, cp.CallPath, is.AllOf(