+++ pkg/bpmn_engine/usertasks_test.go
@@ -3,7 +3,6 @@ package bpmn_engine
 import (
 	"github.com/corbym/gocrest/is"
 	"github.com/corbym/gocrest/then"
-	"github.com/nitram509/lib-bpmn-engine/pkg/spec/BPMN20/process_instance"
 	"testing"
 )
 
@@ -13,11 +12,11 @@ func Test_user_tasks_can_be_handled(t *testing.T) {
 	process, err := bpmnEngine.LoadFromFile("../../test-cases/simple-user-task.bpmn")
 	then.AssertThat(t, err, is.Nil())
 	cp := CallPath{}
-	bpmnEngine.NewTaskHandler().Id("user-task").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Id("user-task").Handler(cp.TaskHandler)
 
 	instance, _ := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
 
-	then.AssertThat(t, instance.State, is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, instance.State, is.EqualTo(Completed))
 	then.AssertThat(t, cp.CallPath, is.EqualTo("user-task"))
 }
 
@@ -33,15 +32,17 @@ func Test_user_tasks_can_be_continue(t *testing.T) {
 	userConfirm := false
 	bpmnEngine.NewTaskHandler().Id("user-task").Handler(func(job ActivatedJob) {
 		if userConfirm {
-			cp.CallPathHandler(job)
+			cp.TaskHandler(job)
 		}
 	})
-	bpmnEngine.RunOrContinueInstance(instance.InstanceKey)
+	_, err = bpmnEngine.RunOrContinueInstance(instance.InstanceKey)
+	then.AssertThat(t, err, is.Nil())
 
 	userConfirm = true
 
-	bpmnEngine.RunOrContinueInstance(instance.InstanceKey)
+	_, err = bpmnEngine.RunOrContinueInstance(instance.InstanceKey)
+	then.AssertThat(t, err, is.Nil())
 
-	then.AssertThat(t, instance.State, is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, instance.State, is.EqualTo(Completed))
 	then.AssertThat(t, cp.CallPath, is.EqualTo("user-task"))
 }