+++ tests/serializing_test.go
@@ -5,7 +5,6 @@ import (
 	"github.com/corbym/gocrest/is"
 	"github.com/corbym/gocrest/then"
 	"github.com/nitram509/lib-bpmn-engine/pkg/bpmn_engine"
-	"github.com/nitram509/lib-bpmn-engine/pkg/spec/BPMN20/process_instance"
 	"os"
 	"testing"
 )
@@ -49,7 +48,44 @@ func Test_Marshal_Unmarshal_Jobs(t *testing.T) {
 	// then
 	instance, err = bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
 	then.AssertThat(t, err, is.Nil())
-	then.AssertThat(t, instance.GetState(), is.EqualTo(process_instance.ACTIVE))
+	then.AssertThat(t, instance.GetState(), is.EqualTo(bpmn_engine.Active))
+}
+
+func Test_Marshal_Unmarshal_partially_executed_jobs_continue_where_left_of_before_marshalling(t *testing.T) {
+	// setup
+	bpmnEngine := bpmn_engine.New()
+	cp := CallPath{}
+	bpmnEngine.NewTaskHandler().Id("id-a-1").Handler(cp.CallPathHandler)
+
+	// given
+	pi, err := bpmnEngine.LoadFromFile("../test-cases/parallel-gateway-flow.bpmn")
+	then.AssertThat(t, err, is.Nil())
+
+	// when
+	instance, err := bpmnEngine.CreateAndRunInstance(pi.ProcessKey, nil)
+	then.AssertThat(t, err, is.Nil())
+	then.AssertThat(t, cp.CallPath, is.EqualTo("id-a-1"))
+
+	instance, err = bpmnEngine.RunOrContinueInstance(instance.InstanceKey)
+	bytes := bpmnEngine.Marshal()
+	then.AssertThat(t, len(bytes), is.GreaterThan(32))
+
+	if EnableJsonDataDump {
+		os.WriteFile("temp.marshal.parallel-jobs.json", bytes, 0644)
+	}
+
+	// when
+	bpmnEngine, err = bpmn_engine.Unmarshal(bytes)
+	bpmnEngine.NewTaskHandler().Id("id-b-1").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Id("id-b-2").Handler(cp.CallPathHandler)
+	then.AssertThat(t, err, is.Nil())
+
+	// then
+	instance, err = bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
+	then.AssertThat(t, err, is.Nil())
+	then.AssertThat(t, instance.GetState(), is.EqualTo(bpmn_engine.Completed))
+	then.AssertThat(t, cp.CallPath, is.EqualTo("id-a-1,id-b-1,id-b-2"))
+
 }
 
 func Test_Marshal_Unmarshal_Remain_Handler(t *testing.T) {
@@ -65,7 +101,7 @@ func Test_Marshal_Unmarshal_Remain_Handler(t *testing.T) {
 	// when
 	instance, err := bpmnEngine.CreateInstance(pi.ProcessKey, nil)
 	then.AssertThat(t, err, is.Nil())
-	then.AssertThat(t, instance.GetState(), is.EqualTo(process_instance.READY))
+	then.AssertThat(t, instance.GetState(), is.EqualTo(bpmn_engine.Ready))
 	bytes := bpmnEngine.Marshal()
 
 	if EnableJsonDataDump {
@@ -80,7 +116,7 @@ func Test_Marshal_Unmarshal_Remain_Handler(t *testing.T) {
 	// then
 	instance, err = newEngine.RunOrContinueInstance(instance.GetInstanceKey())
 	then.AssertThat(t, err, is.Nil())
-	then.AssertThat(t, instance.GetState(), is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, instance.GetState(), is.EqualTo(bpmn_engine.Completed))
 
 	then.AssertThat(t, cp.CallPath, is.EqualTo("id"))
 }